+++ server/server-service/server-service-api/src/main/java/com/gc/irc/server/service/api/IAuthenticationService.java
@@ -33,7 +33,7 @@ public interface IAuthenticationService {
      * @param nickname
      *            New NickName
      */
-    void updateUserNickName(int id, String nickname);
+    void updateUserNickName(long id, String nickname);
 
     /**
      * Change the Password of the User.
@@ -43,7 +43,7 @@ public interface IAuthenticationService {
      * @param password
      *            New Password.
      */
-    void updateUserPasword(int id, String password);
+    void updateUserPasword(long id, String password);
 
     /**
      * Get the designed User.
@@ -52,7 +52,7 @@ public interface IAuthenticationService {
      *            User's id.
      * @return The User if exist. Else null.
      */
-    UserInformations getUser(int id);
+    UserInformations getUser(long id);
 
     /**
      * Log an user.
